requirements.txt
asgiref==3.4.1
decorator==5.0.9
Django==3.2.7
django-fakeredis==0.1.2
django-redis==5.0.0
djangorestframework==3.12.4
fakeredis==1.6.0
pytz==2021.1
redis==3.5.3
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.1
validators==0.18.2
